+++ app/src/main/java/com/shuyun/qapp/bean/ExchangeHistoryBean.java
@@ -18,10 +18,10 @@ public class ExchangeHistoryBean {
      * treasureUserChangeDataList : [{"changeTime":1533017199000,"userAccount":"158****1563","ticketNum":"ser35151w6e51","isMine":1,"isWin":1},{"changeTime":1533174736567,"userAccount":"158****1563","ticketNum":"e4dd41b0a55f4bf69b69397e9dd713e9","isMine":1,"isWin":0},{"changeTime":1533174783574,"userAccount":"158****1563","ticketNum":"94db38edf08942ed9f46c2c96f01aef9","isMine":1,"isWin":1},{"changeTime":1533174895768,"userAccount":"158****1563","ticketNum":"ebcd52a8637a4e69b8aac8774705edb7","isMine":1,"isWin":0},{"changeTime":1533194477599,"userAccount":"158****1563","ticketNum":"fde534da0966405aa0a6fe9f86076ab9","isMine":1,"isWin":0}]
      */
 
-    private int scheduleId;
+    private Long scheduleId;
     private String prizeName;
     private String prizePurpose;
-    private int scheduleStatus;
+    private Long scheduleStatus;
     private String mainPic;
     private List<TreasureUserChangeDataListBean> treasureUserChangeDataList;
 
@@ -33,11 +33,11 @@ public class ExchangeHistoryBean {
         this.mainPic = mainPic;
     }
 
-    public int getScheduleId() {
+    public Long getScheduleId() {
         return scheduleId;
     }
 
-    public void setScheduleId(int scheduleId) {
+    public void setScheduleId(Long scheduleId) {
         this.scheduleId = scheduleId;
     }
 
@@ -57,11 +57,11 @@ public class ExchangeHistoryBean {
         this.prizePurpose = prizePurpose;
     }
 
-    public int getScheduleStatus() {
+    public Long getScheduleStatus() {
         return scheduleStatus;
     }
 
-    public void setScheduleStatus(int scheduleStatus) {
+    public void setScheduleStatus(Long scheduleStatus) {
         this.scheduleStatus = scheduleStatus;
     }
 
@@ -85,8 +85,8 @@ public class ExchangeHistoryBean {
         private long changeTime;
         private String userAccount;
         private String ticketNum;
-        private int isMine;
-        private int isWin;
+        private Long isMine;
+        private Long isWin;
 
         public long getChangeTime() {
             return changeTime;
@@ -112,19 +112,19 @@ public class ExchangeHistoryBean {
             this.ticketNum = ticketNum;
         }
 
-        public int getIsMine() {
+        public Long getIsMine() {
             return isMine;
         }
 
-        public void setIsMine(int isMine) {
+        public void setIsMine(Long isMine) {
             this.isMine = isMine;
         }
 
-        public int getIsWin() {
+        public Long getIsWin() {
             return isWin;
         }
 
-        public void setIsWin(int isWin) {
+        public void setIsWin(Long isWin) {
             this.isWin = isWin;
         }
     }
